backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-supabase
